\name{OB-classes}
\Rdversion{1.1}
\docType{class}

\title{Classes from OB}

\description{ These are methods generated by SWIG for R. These should not generally be used
	outside of ChemmineOB, they are listed here to quite some warnings.}


\alias{$,_p_OpenBabel__AliasData-method}
\alias{$,_p_OpenBabel__AliasData-method}
\alias{$,_p_OpenBabel__DoubleType-method}
\alias{$,_p_OpenBabel__FastSearch-method}
\alias{$,_p_OpenBabel__FastSearchIndexer-method}
\alias{$,_p_OpenBabel__FptIndexHeader-method}
\alias{$,_p_OpenBabel__OBAngle-method}
\alias{$,_p_OpenBabel__OBAngleData-method}
\alias{$,_p_OpenBabel__OBAromaticTyper-method}
\alias{$,_p_OpenBabel__OBAtom-method}
\alias{$,_p_OpenBabel__OBAtomAtomIter-method}
\alias{$,_p_OpenBabel__OBAtomBondIter-method}
\alias{$,_p_OpenBabel__OBAtomClassData-method}
\alias{$,_p_OpenBabel__OBAtomTyper-method}
\alias{$,_p_OpenBabel__OBBase-method}
\alias{$,_p_OpenBabel__OBBitVec-method}
\alias{$,_p_OpenBabel__OBBond-method}
\alias{$,_p_OpenBabel__OBBuilder-method}
\alias{$,_p_OpenBabel__OBChainsParser-method}
\alias{$,_p_OpenBabel__OBChiralData-method}
\alias{$,_p_OpenBabel__OBCommentData-method}
\alias{$,_p_OpenBabel__OBConformerData-method}
\alias{$,_p_OpenBabel__OBConversion-method}
\alias{$,_p_OpenBabel__OBDOSData-method}
\alias{$,_p_OpenBabel__OBDescriptor-method}
\alias{$,_p_OpenBabel__OBElectronicTransitionData-method}
\alias{$,_p_OpenBabel__OBElement-method}
\alias{$,_p_OpenBabel__OBElementTable-method}
\alias{$,_p_OpenBabel__OBError-method}
\alias{$,_p_OpenBabel__OBExternalBond-method}
\alias{$,_p_OpenBabel__OBExternalBondData-method}
\alias{$,_p_OpenBabel__OBFFCalculation2-method}
\alias{$,_p_OpenBabel__OBFFCalculation3-method}
\alias{$,_p_OpenBabel__OBFFCalculation4-method}
\alias{$,_p_OpenBabel__OBFFConstraint-method}
\alias{$,_p_OpenBabel__OBFFConstraints-method}
\alias{$,_p_OpenBabel__OBFFParameter-method}
\alias{$,_p_OpenBabel__OBFingerprint-method}
\alias{$,_p_OpenBabel__OBForceField-method}
\alias{$,_p_OpenBabel__OBFormat-method}
\alias{$,_p_OpenBabel__OBGenericData-method}
\alias{$,_p_OpenBabel__OBGlobalDataBase-method}
\alias{$,_p_OpenBabel__OBGridData-method}
\alias{$,_p_OpenBabel__OBInternalCoord-method}
\alias{$,_p_OpenBabel__OBIsotopeTable-method}
\alias{$,_p_OpenBabel__OBMatrixData-method}
\alias{$,_p_OpenBabel__OBMessageHandler-method}
\alias{$,_p_OpenBabel__OBMol-method}
\alias{$,_p_OpenBabel__OBMolAngleIter-method}
\alias{$,_p_OpenBabel__OBMolAtomBFSIter-method}
\alias{$,_p_OpenBabel__OBMolAtomDFSIter-method}
\alias{$,_p_OpenBabel__OBMolAtomIter-method}
\alias{$,_p_OpenBabel__OBMolBondBFSIter-method}
\alias{$,_p_OpenBabel__OBMolBondIter-method}
\alias{$,_p_OpenBabel__OBMolPairIter-method}
\alias{$,_p_OpenBabel__OBMolRingIter-method}
\alias{$,_p_OpenBabel__OBMolTorsionIter-method}
\alias{$,_p_OpenBabel__OBOp-method}
\alias{$,_p_OpenBabel__OBOrbital-method}
\alias{$,_p_OpenBabel__OBOrbitalData-method}
\alias{$,_p_OpenBabel__OBPairData-method}
\alias{$,_p_OpenBabel__OBPlugin-method}
\alias{$,_p_OpenBabel__OBRTree-method}
\alias{$,_p_OpenBabel__OBRandom-method}
\alias{$,_p_OpenBabel__OBResidue-method}
\alias{$,_p_OpenBabel__OBResidueAtomIter-method}
\alias{$,_p_OpenBabel__OBResidueData-method}
\alias{$,_p_OpenBabel__OBResidueIter-method}
\alias{$,_p_OpenBabel__OBRing-method}
\alias{$,_p_OpenBabel__OBRingData-method}
\alias{$,_p_OpenBabel__OBRingSearch-method}
\alias{$,_p_OpenBabel__OBRingTyper-method}
\alias{$,_p_OpenBabel__OBRotationData-method}
\alias{$,_p_OpenBabel__OBSSMatch-method}
\alias{$,_p_OpenBabel__OBSerialNums-method}
\alias{$,_p_OpenBabel__OBSetData-method}
\alias{$,_p_OpenBabel__OBSmartsMatcher-method}
\alias{$,_p_OpenBabel__OBSmartsPattern-method}
\alias{$,_p_OpenBabel__OBSqrtTbl-method}
\alias{$,_p_OpenBabel__OBStopwatch-method}
\alias{$,_p_OpenBabel__OBSymmetryData-method}
\alias{$,_p_OpenBabel__OBTorsion-method}
\alias{$,_p_OpenBabel__OBTorsionData-method}
\alias{$,_p_OpenBabel__OBTypeTable-method}
\alias{$,_p_OpenBabel__OBUnitCell-method}
\alias{$,_p_OpenBabel__OBVectorData-method}
\alias{$,_p_OpenBabel__OBVibrationData-method}
\alias{$,_p_OpenBabel__OBVirtualBond-method}
\alias{$,_p_OpenBabel__SpaceGroup-method}
\alias{$,_p_OpenBabel__matrix3x3-method}
\alias{$,_p_OpenBabel__transform3d-method}
\alias{$,_p_OpenBabel__vector3-method}
\alias{$,_p_std__pairT_unsigned_int_unsigned_int_t-method}

\alias{$,_p_std__vectorT_OpenBabel__OBBond_std__allocatorT_OpenBabel__OBBond_t_t-method}
\alias{$,_p_std__vectorT_OpenBabel__OBGenericData_p_std__allocatorT_OpenBabel__OBGenericData_p_t_t-method}
\alias{$,_p_std__vectorT_OpenBabel__OBInternalCoord_p_std__allocatorT_OpenBabel__OBInternalCoord_p_t_t-method}
\alias{$,_p_std__vectorT_OpenBabel__OBMol_std__allocatorT_OpenBabel__OBMol_t_t-method}
\alias{$,_p_std__vectorT_OpenBabel__OBResidue_std__allocatorT_OpenBabel__OBResidue_t_t-method}
\alias{$,_p_std__vectorT_OpenBabel__OBRing_p_std__allocatorT_OpenBabel__OBRing_p_t_t-method}
\alias{$,_p_std__vectorT_OpenBabel__OBRing_std__allocatorT_OpenBabel__OBRing_t_t-method}
\alias{$,_p_std__vectorT_OpenBabel__vector3_std__allocatorT_OpenBabel__vector3_t_t-method}
\alias{$,_p_std__vectorT_double_std__allocatorT_double_t_t-method}
\alias{$,_p_std__vectorT_int_std__allocatorT_int_t_t-method}
\alias{$,_p_std__vectorT_std__pairT_unsigned_int_unsigned_int_t_std__allocatorT_std__pairT_unsigned_int_unsigned_int_t_t_t-method}
\alias{$,_p_std__vectorT_std__string_std__allocatorT_std__string_t_t-method}
\alias{$,_p_std__vectorT_std__vectorT_OpenBabel__vector3_std__allocatorT_OpenBabel__vector3_t_t_std__allocatorT_std__vectorT_OpenBabel__vector3_std__allocatorT_OpenBabel__vector3_t_t_t_t-method}
\alias{$,_p_std__vectorT_std__vectorT_int_std__allocatorT_int_t_t_std__allocatorT_std__vectorT_int_std__allocatorT_int_t_t_t_t-method}
\alias{$,_p_std__vectorT_std__vectorT_std__pairT_unsigned_int_unsigned_int_t_std__allocatorT_std__pairT_unsigned_int_unsigned_int_t_t_t_std__allocatorT_std__vectorT_std__pairT_unsigned_int_unsigned_int_t_std__allocatorT_std__pairT_unsigned_int_unsigned_int_t_t_t_t_t-method}
\alias{$,_p_std__vectorT_unsigned_int_std__allocatorT_unsigned_int_t_t-method}
\alias{$,_p_stringp-method}
\alias{$<-,_p_OpenBabel__DoubleType-method}
\alias{$<-,_p_OpenBabel__FptIndexHeader-method}
\alias{$<-,_p_OpenBabel__OBAtom-method}
\alias{$<-,_p_OpenBabel__OBAtomAtomIter-method}
\alias{$<-,_p_OpenBabel__OBAtomBondIter-method}
\alias{$<-,_p_OpenBabel__OBBond-method}
\alias{$<-,_p_OpenBabel__OBFFCalculation2-method}
\alias{$<-,_p_OpenBabel__OBFFCalculation3-method}
\alias{$<-,_p_OpenBabel__OBFFCalculation4-method}
\alias{$<-,_p_OpenBabel__OBFFConstraint-method}
\alias{$<-,_p_OpenBabel__OBFFParameter-method}
\alias{$<-,_p_OpenBabel__OBInternalCoord-method}
\alias{$<-,_p_OpenBabel__OBMolAtomBFSIter-method}
\alias{$<-,_p_OpenBabel__OBMolAtomDFSIter-method}
\alias{$<-,_p_OpenBabel__OBMolAtomIter-method}
\alias{$<-,_p_OpenBabel__OBMolBondBFSIter-method}
\alias{$<-,_p_OpenBabel__OBMolBondIter-method}
\alias{$<-,_p_OpenBabel__OBMolRingIter-method}
\alias{$<-,_p_OpenBabel__OBResidueAtomIter-method}
\alias{$<-,_p_OpenBabel__OBRing-method}
\alias{$<-,_p_OpenBabel__OBUnitCell-method}
\alias{$<-,_p_std__pairT_unsigned_int_unsigned_int_t-method}
\alias{[,ExternalReference-method}
\alias{[<-,ExternalReference-method}
\alias{[[<-,_p_OpenBabel__DoubleType,character-method}
\alias{[[<-,_p_OpenBabel__FptIndexHeader,character-method}
\alias{[[<-,_p_OpenBabel__OBAtom,character-method}
\alias{[[<-,_p_OpenBabel__OBAtomAtomIter,character-method}
\alias{[[<-,_p_OpenBabel__OBAtomBondIter,character-method}
\alias{[[<-,_p_OpenBabel__OBBond,character-method}
\alias{[[<-,_p_OpenBabel__OBFFCalculation2,character-method}
\alias{[[<-,_p_OpenBabel__OBFFCalculation3,character-method}
\alias{[[<-,_p_OpenBabel__OBFFCalculation4,character-method}
\alias{[[<-,_p_OpenBabel__OBFFConstraint,character-method}
\alias{[[<-,_p_OpenBabel__OBFFParameter,character-method}
\alias{[[<-,_p_OpenBabel__OBInternalCoord,character-method}
\alias{[[<-,_p_OpenBabel__OBMolAtomBFSIter,character-method}
\alias{[[<-,_p_OpenBabel__OBMolAtomDFSIter,character-method}
\alias{[[<-,_p_OpenBabel__OBMolAtomIter,character-method}
\alias{[[<-,_p_OpenBabel__OBMolBondBFSIter,character-method}
\alias{[[<-,_p_OpenBabel__OBMolBondIter,character-method}
\alias{[[<-,_p_OpenBabel__OBMolRingIter,character-method}
\alias{[[<-,_p_OpenBabel__OBResidueAtomIter,character-method}
\alias{[[<-,_p_OpenBabel__OBRing,character-method}
\alias{[[<-,_p_OpenBabel__OBUnitCell,character-method}
\alias{[[<-,_p_std__pairT_unsigned_int_unsigned_int_t,character-method}
\alias{length,SWIGArray-method}
